Add ds store file to gitignore and made a navigation module to be called inside index js
